Remove blockmap considerations from index cost estimation

+++ src/hnsw.h
@@ -10,7 +10,7 @@
 
 #define LANTERN_INTERNAL_SCHEMA_NAME "_lantern_internal"
 
-static const int INVALID_ELEMENT_LABEL = 0;
+static const unsigned long int INVALID_ELEMENT_LABEL = 0;
 
 #if PG_VERSION_NUM < 110000
 #error "Requires PostgreSQL 11+"
